fix(tauri/ui): remove modal to confirm removing order
